做收藏页面做到一半，逻辑出了问题，正在思考解决中，但晚上要回寝室了，明天再解决。

+++ app/src/main/java/com/hnqcgc/redfirecookbook/logic/db/RedFireCookBookDB.java
@@ -28,7 +28,7 @@ public class RedFireCookBookDB {
         return collectionDao.loadAllCollection();
     }
 
-    public int deleteCollectionById(int id) {
+    public long deleteCollectionById(long id) {
         return collectionDao.deleteCollectionById(id);
     }
